requirements.txt
bleach==6.2.0
deepspeed==0.17.1
einops==0.8.1
flash_attn==2.8.0.post2
fvcore==0.1.5.post20221221
gradio==5.34.2
habitat_sim==0.2.4.dev20230403
huggingface_hub==0.33.0
lxml==5.4.0
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
peft==0.15.2
Pillow==11.2.1
pycocotools==2.0.10
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.4
rerun==1.0.30
scipy==1.15.3
simple_3dviz==0.7.0
skimage==0.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
trimesh==4.6.12
